requirements.txt
numpy
pymsteams
pymupdf
pytesseract
pytest
python-dotenv
rapidfuzz
requests